requirements.txt
streamlit>=1.24.0
plotly>=5.14.0
Pillow>=9.5.0
watchdog>=3.0.0
